extract bio and projects components
